\name{CorporaCorpus-package}
\alias{CorporaCorpus-package}
\alias{CorporaCorpus}
\docType{package}

\title{A collection of small corpora prepared by the Centre for Corpus Research}
\description{
    This package is a container for the following corpora
    \tabular{ll}{
        Name \tab Description \cr
        \code{\link{dickens}} \tab Dickens' novels \cr
        \code{\link{ncnb}} \tab 19th Century Novels (British)
    }
}
\details{
    The corpora are prepared in such a manner as to simplify analysis using R.
    Details of editing and homogenization of typesetting can be found in the
    \href{../doc/faq.pdf}{FAQ} vignette.

    The content of each corpus is documented in detail in the
    \href{../doc/corpora.pdf}{Corpora} vignette.

    For a list of all documentation use \code{library(help="CorporaCorpus")}.
}

\author{
  Maintainer: Anthony Hennessey <ah@ravingmantis.com>.
}
